requirements.txt
django==3.2.5
django-cors-headers==3.5.0
djangorestframework==3.12.4
#psycopg2>=2.7.5, <2.8.0
psycopg2
#gunicorn==20.0.4
drf-yasg